requirements.txt
asgiref==3.2.3
Django==3.0.2
django-appconf==1.0.3
django-ckeditor==5.4.0
django-imagekit==4.0.2
django-js-asset==1.2.2
django-model-utils==4.0.0
django-notifications-hq==1.6.0
django-simpleui==3.9.1
dnspython==1.16.0
jsonfield==3.1.0
mysql-connector-python==8.0.19
mysqlclient==1.4.6
pilkit==2.0
Pillow==7.0.0
protobuf==3.6.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
swapper==1.1.2.post1
